+++ src/main/java/com/propiconnect/inmoshare/properties/domain/model/commands/CreatePropertyCommand.java
@@ -1,4 +1,4 @@
 package com.propiconnect.inmoshare.properties.domain.model.commands;
 
-public record CreatePropertyCommand(String ownerName,Long ownerId,String city,String type, String address, String description, String propertyType, String rentalType, String image, Long initialPrice) {
+public record CreatePropertyCommand(String ownerName,Long ownerId,String city,String type, String address, String description, String propertyType, String rentalType, String image, Double initialPrice) {
 }
